+++ src/ORM/Query/AST/ContainsFunction.php
@@ -1,0 +1,43 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vrok\DoctrineAddons\ORM\Query\AST;
+
+use Doctrine\ORM\Query\AST\ArithmeticExpression;
+use Doctrine\ORM\Query\AST\Functions\FunctionNode;
+use Doctrine\ORM\Query\Lexer;
+use Doctrine\ORM\Query\Parser;
+use Doctrine\ORM\Query\SqlWalker;
+
+/**
+ * Implements support for the postgres-only "@>" operator on jsonb fields or
+ * arrays, to search for elements.
+ *
+ * @see https://www.postgresql.org/docs/current/functions-array.html
+ * @see https://www.postgresql.org/docs/current/datatype-json.html#JSON-CONTAINMENT
+ */
+class ContainsFunction extends FunctionNode
+{
+    private ArithmeticExpression $expr1;
+    private ArithmeticExpression $expr2;
+
+    public function parse(Parser $parser): void
+    {
+        $parser->match(Lexer::T_IDENTIFIER);
+        $parser->match(Lexer::T_OPEN_PARENTHESIS);
+        $this->expr1 = $parser->ArithmeticExpression();
+        $parser->match(Lexer::T_COMMA);
+        $this->expr2 = $parser->ArithmeticExpression();
+        $parser->match(Lexer::T_CLOSE_PARENTHESIS);
+    }
+
+    public function getSql(SqlWalker $sqlWalker): string
+    {
+        return sprintf(
+            '(%s @> %s)',
+            $this->expr1->dispatch($sqlWalker),
+            $this->expr2->dispatch($sqlWalker)
+        );
+    }
+}